implement player_controller.hpp add actors to the world add resources load manager add display logger
player_actor_moving.cpp tests add
3types_tiles.map

+++ include/TopDownEngine/core/core.hpp
@@ -9,7 +9,7 @@
 #include <array>
 
 namespace Engine {
-    using uint = uint32_t;
+    using uint = uint64_t;
 
     /// Class with tick function that called every frame
     class Tickable {
